+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- rewrite code to TypeScript
+
+### Fixed
+- fix test for `fetchSilkeyPublicKey()`
+
 ## [0.3.2] - 2020-11-26
 ### Added
 - `generateSSORequestParams` throws on empty PK