+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.3] - 2020-12-05
+### Removed
+- remove default export
+
 ## [0.4.2] - 2020-12-04
 ### Fixed
 - make abi part of the source code, so we don't need to use `fs` as it will not work for React